src/test/java/dev/chernykh/cellular/user/UserControllerTest.java
package dev.chernykh.cellular.user;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static java.util.Arrays.asList;
import static org.hamcrest.Matchers.hasSize;
import static org.hamcrest.Matchers.is;
import static org.mockito.Mockito.*;
import static org.springframework.http.MediaType.APPLICATION_JSON;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringRunner.class)
@WebMvcTest(UserController.class)
public class UserControllerTest {
    @Autowired
    private MockMvc mvc;
    @MockBean
    private UserRepository repository;
    
    @Test
    public void shouldReturnListOfUsers() throws Exception {
        when(repository.findAll()).thenReturn(asList(
            new User(1, "John Doe", 1),
            new User(7, "James Bond", 7)
        ));
        
        mvc.perform(get("/api/users").accept(APPLICATION_JSON))
            .andExpect(status().isOk())
            .andExpect(content().contentTypeCompatibleWith(APPLICATION_JSON))
            .andExpect(jsonPath("$", hasSize(2)))
            .andExpect(jsonPath("$[0].id", is(1)))
            .andExpect(jsonPath("$[0].fullName", is("John Doe")))
            .andExpect(jsonPath("$[0].id", is(1)))
            .andExpect(jsonPath("$[1].id", is(7)))
            .andExpect(jsonPath("$[1].fullName", is("James Bond")))
            .andExpect(jsonPath("$[1].id", is(7)));
        
        verify(repository, times(1)).findAll();
    }
}
